ignux-backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.4.0
alembic==1.13.0
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
attrs==25.4.0
backports-datetime-fromisoformat==2.0.3
bcrypt==5.0.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
environs==14.5.0
exceptiongroup==1.3.1
fastapi==0.104.1
fastapi_cors==0.0.6
frozenlist==1.8.0
greenlet==3.3.0
gunicorn==21.2.0
h11==0.16.0
httptools==0.7.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.1.1
multidict==6.7.0
packaging==25.0
passlib==1.7.4
propcache==0.4.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.23
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.15.0
uvicorn==0.24.0
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
